requirement.txt
streamlit
pandas
faiss-cpu
sentence-transformers
numpy
google-generativeai
